more 64

+++ src/backend/debug.c
@@ -107,7 +107,7 @@ void WRTYxx(tym_t t)
 #endif
     t = tybasic(t);
     if (t >= TYMAX)
-    {   dbg_printf("TY %lx\n",t);
+    {   dbg_printf("TY %lx\n",(long)t);
         assert(0);
     }
     dbg_printf("TY%s ",tystring[tybasic(t)]);
@@ -185,7 +185,7 @@ void WReqn(elem *e)
         ferr(" ");
         WROP(e->Eoper);
         if (e->Eoper == OPstreq)
-            dbg_printf("%ld",e->Enumbytes);
+            dbg_printf("%ld",(long)e->Enumbytes);
         ferr(" ");
         if (OTbinary(e->E2->Eoper))
         {       nest++;
@@ -222,7 +222,7 @@ void WReqn(elem *e)
                 if (e->EV.sp.Vsym->Ssymnum != -1)
                     dbg_printf("(%d)",e->EV.sp.Vsym->Ssymnum);
                 if (e->Eoffset != 0)
-                        dbg_printf(".%ld",e->Eoffset);
+                        dbg_printf(".%ld",(long)e->Eoffset);
                 break;
             case OPasm:
 #if TARGET_MAC
@@ -238,7 +238,7 @@ void WReqn(elem *e)
             case OPstring:
                 dbg_printf("\"%s\"",e->EV.ss.Vstring);
                 if (e->EV.ss.Voffset)
-                    dbg_printf("+%ld",e->EV.ss.Voffset);
+                    dbg_printf("+%ld",(long)e->EV.ss.Voffset);
                 break;
             case OPmark:
             case OPgot:
